# swagger.yaml
openapi: 3.0.3
info:
  title: Camping API
  version: 1.1.0
  contact: {}
servers:
  - url: "http://localhost:8080/api"
paths:
...
  /feedback/create:
    post:
      tags:
        - Feedback
      summary: Create feedback
      description: ""
      operationId: createFeedback
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  example: RgYM60evCyfBxbYmxWg7o0Pqhpz2
                message:
                  type: string
                  example: Write your feedback here
            example:
              userId: RgYM60evCyfBxbYmxWg7o0Pqhpz2
              message: Write your feedback here
      responses:
        "500":
          description: Internal Server Error
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Feedback"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Validation error some field is missing
...
components:
  schemas:
...
    Feedback:
      type: object
      properties:
        _id:
          type: string
          description: Feedback unique id (autogenerated)
        userId:
          type: string
          description: User uid
        message:
          type: string
          description: Feedback message
        createdAt:
          type: string
          format: date
          description: Feedback datetime UTC format (autogenerated)
      required: [userId, message]
...